Move code generation from ndau pkg
Per the README: it's tedious to put code generation in the same repo as
its generated code, because go wants to compile the whole repo at once,
so any misgeneration prevented iteration until it was fixed. That was a
pain. We therefore moved it here and also in the process fixed the
SignableBytes implementation so that it now generates code which
compiles and which passes the ndau unit tests.